+++ XRViewer/AppDelegate.swift
@@ -1,4 +1,5 @@
 import UIKit
+import CocoaLumberjack
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {