+++ XRViewer/Resources/Prefix.swift
@@ -1,0 +1,2 @@
+let REQUESTED_URL_KEY = "requestedURL"
+let PREFER_FPS = 60