+++ XRViewer/Resources/XRViewer-Bridging-Header.h
@@ -3,9 +3,15 @@
 //
 
 #import "ARKHelper.h"
-#import "Prefix.h"
 #import "ARKController.h"
 #import "WebARKHeader.h"
+#import <CocoaLumberjack/CocoaLumberjack.h>
+
+#if DEBUG
+static const DDLogLevel ddLogLevel = DDLogLevelVerbose;
+#else
+static const DDLogLevel ddLogLevel = DDLogLevelWarning;
+#endif
 
 typedef NS_ENUM(NSInteger, ResetTrackingOption) {
     ResetTracking,