+++ XRViewer/ViewController/ViewController+ARSCNViewDelegate.swift
@@ -1,4 +1,5 @@
 import ARKit
+import CocoaLumberjack
 
 extension ViewController: ARSCNViewDelegate {